In Progress: handle api versions

+++ src/ErrorCodes.cs
@@ -4,7 +4,7 @@ namespace codecrafters;
 
 public static class ErrorCodes
 {
-    public const int UNKNOWN_SERVER_ERROR = -1;
-    public const int NONE = 0;
-    public const int UNSUPPORTED_VERSION = 35;
+    public const short UNKNOWN_SERVER_ERROR = -1;
+    public const short NONE = 0;
+    public const short UNSUPPORTED_VERSION = 35;
 }
